% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_cache.R
\name{caf_create}
\alias{caf_create}
\title{Create cache-aware function}
\usage{
caf_create(fun = NULL, observes = list(), .refresh_default = TRUE,
  .verbose_default = FALSE)
}
\description{
Create cache-aware function
}
